pyproject.toml
[project]
name = "sentient-artifacts-tui"
version = "0.1.0"
description = "Textual TUI for the Sentient Artifacts bot swarm"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "textual>=0.52.0",
    "textual-image>=0.8.4",
    "pillow>=12.1.0",
    "httpx>=0.26.0",
    "websockets>=16.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
sentient-tui = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sentient_artifacts", "main.py"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["B", "D", "E", "F", "I", "N", "UP"]
ignore = ["E501", "D203", "D205", "D212", "D400", "D401", "D415"]

[tool.ruff.lint.pydocstyle]
convention = "google"
